fix comment rp hw5_v3
