+++ app/src/main/java/ru/thstdio/aa2020/api/response/JsonGenre.kt
@@ -6,4 +6,4 @@ import ru.thstdio.aa2020.data.Genre
 @Serializable
 class JsonGenre(val id: Long, val name: String)
 
-fun JsonGenre.toGenre() = Genre(id = this.id, name = this.name)
+fun JsonGenre.toGenre(): Genre = Genre(id = this.id, name = this.name)
