+++ app/src/main/java/ru/thstdio/aa2020/ui/list/FragmentMoviesList.kt
@@ -1,6 +1,7 @@
 package ru.thstdio.aa2020.ui.list
 
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -17,20 +18,16 @@ class FragmentMoviesList : FragmentNavigation(R.layout.fragment_movies_list) {
     }
 
     private val binding: FragmentMoviesListBinding by viewBinding()
-    private var repository: Repository? = null
     private val exceptionHandler = CoroutineExceptionHandler { coroutineContext, exception ->
-        println("CoroutineExceptionHandler got $exception in $coroutineContext")
+        Log.e("Error", "CoroutineExceptionHandler got $exception in $coroutineContext")
     }
     private var scope = CoroutineScope(
-        Dispatchers.Main.immediate + exceptionHandler
+        SupervisorJob() +
+                Dispatchers.Main.immediate + exceptionHandler
     )
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        scope = CoroutineScope(
-            SupervisorJob() +
-                    Dispatchers.Main.immediate + exceptionHandler
-        )
         binding.recyclerView.layoutManager =
             GridLayoutManager(context, 2, RecyclerView.VERTICAL, false)
         loadMoviesList()
@@ -38,8 +35,9 @@ class FragmentMoviesList : FragmentNavigation(R.layout.fragment_movies_list) {
     }
 
     private fun loadMoviesList() {
+        val repository: Repository = Repository(requireActivity().applicationContext)
         scope.launch {
-            repository?.downloadMovies()?.let { listMovies ->
+            repository.downloadMovies().let { listMovies ->
                 binding.recyclerView.adapter =
                     CinemaListAdapter(listMovies) { cinema -> router.openDetail(cinema.id) }
             }
